Task 3 - Remote monitoring
In here I used a python code to read data from serial monitor, and stored them in a postgresql data base. and then i wrote a code to get those data from data base and visualize them in a web based dashboard. I used Next js and MUI for it.